+++ Tests/SwiftNavigationTests/ObserveTests.swift
@@ -1,10 +1,10 @@
-#if swift(>=6)
-  import SwiftNavigation
-  import XCTest
+import SwiftNavigation
+import XCTest
 
-  class ObserveTests: XCTestCase {
+class ObserveTests: XCTestCase {
+  #if swift(>=6)
     @MainActor
-    func testCompiles() {
+    func testIsolation() {
       var count = 0
       let token = SwiftNavigation.observe {
         count = 1
@@ -12,5 +12,17 @@
       XCTAssertEqual(count, 1)
       _ = token
     }
+  #endif
+
+  @MainActor
+  func testTokenStorage() {
+    var count = 0
+    observe {
+      count += 1
+    }
+    observe {
+      count += 1
+    }
+    XCTAssertEqual(count, 2)
   }
-#endif
+}
